refactor harvest module into a bunch of modules
