+++ eagleproject/notify/views.py
@@ -1,4 +1,3 @@
-import sys
 from datetime import datetime, timedelta
 from django.http import HttpResponse
